containers/user/sources/requirements.txt
numpy
psutil
python-dotenv
opencv-python
six
docker
